Scalable changes applied for file import
Instead of hardcoding each individual list, the changes applied allow
for the individual to import as many stocks / indexes as they would
like.

These changes were not reflected in the bloomberg class, but in the
timeSeriesPackage.py file instead. This was to prevent from changing
data that may have been specifically needed for the Bloomberg class.