Performance_DD (*DD - Double Dictionary)
Able to get the majority of the code to work.

The problem is mainly that the dictionary method does not recognize the
Real Total Return Values as floats, but as strings. Therefore, I cannot
run any calculations on them. I’ve tried casting, as you can see, but
we need to somehow cast in a way that allows us to divide.